fix: getByIdMiddleware refactor
