+++ server/controller.js
@@ -1,5 +1,5 @@
 const Business = require("./models/business");
-const Category = require("./models/categories");
+const Category = require("./models/category");
 
 exports.getAllMiddleware = (modelName) => {
   return async (_req, res) => {
@@ -12,26 +12,16 @@ exports.getAllMiddleware = (modelName) => {
   };
 };
 
-exports.getBusinessById = async (req, res) => {
-  try {
-    const business = await Business.findById(req.params.id);
-    if (!business) {
-      return res.status(404).json({ message: "Business not found" });
+exports.getByIdMiddleware = (modelName) => {
+  return async (req, res) => {
+    try {
+      const result = await modelName.findById(req.params.id);
+      if (!result) {
+        return res.status(404).json({ message: "Result not found" });
+      }
+      res.status(200).json(result);
+    } catch (error) {
+      res.status(500).json({ message: error.message });
     }
-    res.status(200).json(business);
-  } catch (error) {
-    res.status(500).json({ message: error.message });
-  }
-};
-
-exports.getCategoryById = async (req, res) => {
-  try {
-    const category = await Category.findById(req.params.id);
-    if (!category) {
-      return res.status(404).json({ message: "Category not found" });
-    }
-    res.status(200).json(category);
-  } catch (error) {
-    res.status(500).json({ message: error.message });
   }
 };
